+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.1 (Unreleased)
 
-- Add github ci workflow [[GH-10](https://github.com/umbracle/eth2-validator/issues/10)]
+- feat: Add volume storage [[GH-1](https://github.com/umbracle/eth2-validator/issues/1)]
+- feat: Add github ci workflow [[GH-10](https://github.com/umbracle/eth2-validator/issues/10)]
 
 # 0.1.0 (12 May, 2022)